requirements.txt
python-dotenv
Faker
psycopg2